refactor: create registration hook
